@@ -128,4 +128,4 @@ draw2d.util.JSON = {
         }
 };
 
-module.exports = draw2d.util.JSON;
+export default draw2d.util.JSON;
