@@ -97,12 +97,11 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
       // again.
       this.isMoving = false
 
-      let _this = this
-      this.moveListener = function (figure) {
-        if (figure === _this.sourcePort) {
-          _this.setStartPoint(_this.sourcePort.getAbsoluteX(), _this.sourcePort.getAbsoluteY())
+      this.moveListener = (figure) => {
+        if (figure === this.sourcePort) {
+          this.setStartPoint(this.sourcePort.getAbsoluteX(), this.sourcePort.getAbsoluteY())
         } else {
-          _this.setEndPoint(_this.targetPort.getAbsoluteX(), _this.targetPort.getAbsoluteY())
+          this.setEndPoint(this.targetPort.getAbsoluteX(), this.targetPort.getAbsoluteY())
         }
       }
 
@@ -140,9 +139,7 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
 
         // fire the events to all listener
         this.sourcePort.fireEvent("disconnect", {port: this.sourcePort, connection: this})
-        if (this.canvas !== null) {
-          this.canvas.fireEvent("disconnect", {"port": this.sourcePort, "connection": this})
-        }
+        this.canvas?.fireEvent("disconnect", {"port": this.sourcePort, "connection": this})
         this.sourcePort.onDisconnect(this)
 
         this.fireSourcePortRouteEvent()
@@ -154,9 +151,7 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
 
         // fire the events to all listener
         this.targetPort.fireEvent("disconnect", {port: this.targetPort, connection: this})
-        if (this.canvas !== null) {
-          this.canvas.fireEvent("disconnect", {"port": this.targetPort, "connection": this})
-        }
+        this.canvas?.fireEvent("disconnect", {"port": this.targetPort, "connection": this})
         this.targetPort.onDisconnect(this)
 
         this.fireTargetPortRouteEvent()
@@ -174,9 +169,7 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
 
         // fire the events to all listener
         this.sourcePort.fireEvent("connect", {port: this.sourcePort, connection: this})
-        if (this.canvas !== null) {
-          this.canvas.fireEvent("connect", {"port": this.sourcePort, "connection": this})
-        }
+        this.canvas?.fireEvent("connect", {"port": this.sourcePort, "connection": this})
         this.sourcePort.onConnect(this)
 
         this.fireSourcePortRouteEvent()
@@ -188,9 +181,7 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
 
         // fire the events to all listener
         this.targetPort.fireEvent("connect", {port: this.targetPort, connection: this})
-        if (this.canvas !== null) {
-          this.canvas.fireEvent("connect", {"port": this.targetPort, "connection": this})
-        }
+        this.canvas?.fireEvent("connect", {"port": this.targetPort, "connection": this})
         this.targetPort.onConnect(this)
 
         this.fireTargetPortRouteEvent()
@@ -240,16 +231,16 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
      * @returns {this}
      **/
     setSourceDecorator: function (decorator) {
-      if(this.sourceDecorator){
-        this.sourceDecorator.setParent(null)
-      }
+
+      this.sourceDecorator?.setParent(null)
+
       this.sourceDecorator = decorator
       this.routingRequired = true
       this.sourceDecorator.setParent(this)
-      if (this.sourceDecoratorNode !== null) {
-        this.sourceDecoratorNode.remove()
-        this.sourceDecoratorNode = null
-      }
+
+      this.sourceDecoratorNode?.remove()
+      this.sourceDecoratorNode = null
+
       this.repaint()
 
       return this
@@ -273,16 +264,15 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
      * @returns {this}
      **/
     setTargetDecorator: function (decorator) {
-      if(this.targetDecorator){
-        this.targetDecorator.setParent(null)
-      }
+      this.targetDecorator?.setParent(null)
+
       this.targetDecorator = decorator
       this.routingRequired = true
       this.targetDecorator.setParent(this)
-      if (this.targetDecoratorNode !== null) {
-        this.targetDecoratorNode.remove()
-        this.targetDecoratorNode = null
-      }
+
+      this.targetDecoratorNode?.remove()
+      this.targetDecoratorNode = null
+
       this.repaint()
 
       return this
@@ -503,12 +493,8 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
       this._super(figure)
 
       if (this.shape !== null) {
-        if (this.targetDecoratorNode !== null) {
-          this.targetDecoratorNode.insertAfter(this.shape)
-        }
-        if (this.sourceDecoratorNode !== null) {
-          this.sourceDecoratorNode.insertAfter(this.shape)
-        }
+        this.targetDecoratorNode?.insertAfter(this.shape)
+        this.sourceDecoratorNode?.insertAfter(this.shape)
       }
 
       return this
@@ -575,9 +561,8 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
         this.sourcePort.fireEvent("disconnect", {port: this.sourcePort, connection: this})
         // it is possible that a connection has already a port but is not assigned to
         // a canvas. In this case we must check if the canvas set correct before we fire this event
-        if (this.canvas !== null) {
-          this.canvas.fireEvent("disconnect", {"port": this.sourcePort, "connection": this})
-        }
+        this.canvas?.fireEvent("disconnect", {"port": this.sourcePort, "connection": this})
+
         this.sourcePort.onDisconnect(this)
       }
 
@@ -590,9 +575,9 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
       this.fireSourcePortRouteEvent()
       this.sourcePort.connections.add(this)
       this.sourcePort.on("move", this.moveListener)
-      if (this.canvas !== null) {
-        this.canvas.fireEvent("connect", {"port": this.sourcePort, "connection": this})
-      }
+
+      this.canvas?.fireEvent("connect", {"port": this.sourcePort, "connection": this})
+
       this.sourcePort.fireEvent("connect", {port: this.sourcePort, connection: this})
       this.sourcePort.onConnect(this)
 
@@ -623,9 +608,8 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
         this.targetPort.fireEvent("disconnect", {port: this.targetPort, connection: this})
         // it is possible that a connection has already a port but is not assigned to
         // a canvas. In this case we must check if the canvas set correct before we fire this event
-        if (this.canvas !== null) {
-          this.canvas.fireEvent("disconnect", {"port": this.targetPort, "connection": this})
-        }
+        this.canvas?.fireEvent("disconnect", {"port": this.targetPort, "connection": this})
+
         this.targetPort.onDisconnect(this)
       }
 
@@ -638,9 +622,7 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
       this.fireTargetPortRouteEvent()
       this.targetPort.connections.add(this)
       this.targetPort.on("move", this.moveListener)
-      if (this.canvas !== null) {
-        this.canvas.fireEvent("connect", {"port": this.targetPort, "connection": this})
-      }
+      this.canvas?.fireEvent("connect", {"port": this.targetPort, "connection": this})
       this.targetPort.fireEvent("connect", {port: this.targetPort, connection: this})
       this.targetPort.onConnect(this)
 
@@ -659,6 +641,21 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
     },
 
     /**
+     * Returns the peer port of the connection for a given port or nulll if `port` not part of this connection
+     * @param {draw2d.Port} port 
+     * @returns  {draw2d.Port}
+     */
+    getPeerPort: function(port) {
+      if(port === this.sourcePort){
+        return this.targetPort
+      }
+      if(port === this.targetPort){
+        return this.sourcePort
+      }
+      return null
+    },
+
+    /**
      *
      * Method returns true if the connection has at least one common draw2d.Port with the given connection.
      *
@@ -668,9 +665,9 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
      */
     sharingPorts: function (other) {
       return this.sourcePort == other.sourcePort ||
-        this.sourcePort == other.targetPort ||
-        this.targetPort == other.sourcePort ||
-        this.targetPort == other.targetPort
+             this.sourcePort == other.targetPort ||
+             this.targetPort == other.sourcePort ||
+             this.targetPort == other.targetPort
     },
 
 
@@ -689,17 +686,11 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
 
       this._super(canvas)
 
+      this.sourceDecoratorNode?.remove()
+      this.sourceDecoratorNode = null
 
-
-      if (this.sourceDecoratorNode !== null) {
-        this.sourceDecoratorNode.remove()
-        this.sourceDecoratorNode = null
-      }
-
-      if (this.targetDecoratorNode !== null) {
-        this.targetDecoratorNode.remove()
-        this.targetDecoratorNode = null
-      }
+      this.targetDecoratorNode?.remove()
+      this.targetDecoratorNode = null
 
       if (this.canvas === null) {
         if (this.sourcePort !== null) {
@@ -796,7 +787,7 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
      * @private
      **/
     fireSourcePortRouteEvent: function () {
-      this.sourcePort.getConnections().each(function (i, conn) {
+      this.sourcePort.getConnections().each((i, conn) => {
         conn.routingRequired = true
         conn.repaint()
       })
@@ -809,7 +800,7 @@ draw2d.Connection = draw2d.shape.basic.PolyLine.extend(
       // enforce a repaint of all connections which are related to this port
       // this is required for a "FanConnectionRouter" or "ShortesPathConnectionRouter"
       //
-      this.targetPort.getConnections().each(function (i, conn) {
+      this.targetPort.getConnections().each((i, conn) => {
         conn.routingRequired = true
         conn.repaint()
       })
