@@ -48,6 +48,7 @@ draw2d.policy.canvas.DropInterceptorPolicy = draw2d.policy.canvas.CanvasPolicy.e
    * @returns {draw2d.Figure} the calculated connect intent or <b>null</b> if the interceptor uses the veto right
    */
   delegateTarget: function (connectInquirer, connectIntent) {
+
     // a composite accept any kind of figures exceptional ports
     //
     if (!(connectInquirer instanceof draw2d.Port) && connectIntent instanceof draw2d.shape.composite.StrongComposite) {
@@ -68,9 +69,13 @@ draw2d.policy.canvas.DropInterceptorPolicy = draw2d.policy.canvas.CanvasPolicy.e
       return null
     }
 
-    // consider the max possible connections for this port
-    //
-    if (connectIntent.getConnections().getSize() >= connectIntent.getMaxFanOut()) {
+    // ////////////////////////////////////////////////////////////
+    // at this stage we a sure, that we have two ports as argument
+    // ////////////////////////////////////////////////////////////
+
+    // It is not possible to create a loop back connection at the moment.
+    // Reason: no connection router implemented for this case
+    if (connectInquirer.getParent() === connectIntent.getParent()) {
       return null
     }
 
@@ -84,23 +89,31 @@ draw2d.policy.canvas.DropInterceptorPolicy = draw2d.policy.canvas.CanvasPolicy.e
       return null
     }
 
-    // It is not possible to create a loop back connection at the moment.
-    // Reason: no connection router implemented for this case
-    if ((connectInquirer instanceof draw2d.Port) && (connectIntent instanceof draw2d.Port)) {
-      if (connectInquirer.getParent() === connectIntent.getParent()) {
-        return null
-      }
-    }
 
     // redirect the dragEnter handling to the hybrid port
     //
-    if ((connectInquirer instanceof draw2d.Port) && (connectIntent instanceof draw2d.shape.node.Hub)) {
-      return connectIntent.getHybridPort(0)
-    }
+    if ((connectIntent.getParent() instanceof draw2d.shape.node.Hub)) {
 
-    // return the connectTarget determined by the framework or delegate it to another
-    // figure.
+      let inputConnections = connectIntent.getConnections().asArray().filter( con => (con.getPeerPort(connectIntent) instanceof draw2d.OutputPort))
+      let outputConnections = connectIntent.getConnections().asArray().filter( con => (con.getPeerPort(connectIntent) instanceof draw2d.InputPort))
+
+      if (connectInquirer instanceof draw2d.InputPort && outputConnections.length >= connectIntent.getMaxFanOut()) {
+        return null
+      }
+
+      if (connectInquirer instanceof draw2d.OutputPort && inputConnections.length >= connectIntent.getMaxFanIn()) {
+        return null
+      }
+
+      if (connectInquirer instanceof draw2d.HybridPort && inputConnections.length >= connectIntent.getMaxFanIn()) {
+        return null
+      }
+
+      return connectIntent
+    } 
+
     return connectIntent
+
   }
 
 })
